+++ test.sh
@@ -5,4 +5,4 @@ set -e
 rm -rf .box-venv
 mkdir .box-venv
 
-sandbox -v $(pwd):/pwd -v ~/.mitmproxy:/home/sprrw/.mitmproxy -v $(pwd)/.box-venv:/pwd/.venv -- bash -c 'cd /pwd; exec bash'
+sandbox -v $(pwd):/pwd -v ~/.mitmproxy:/home/sprrw/.mitmproxy -v $(pwd)/.box-venv:/pwd/.venv --reset-on-done -- bash -c 'cd /pwd; exec bash'